Master -> Prime [1] [03/05/2023] (#93)
* fix diona mind runtime (#57)

* Tweak: solonuke nerf (#56)

* Upstream Merge [1] 29/04/2023 (#60)

* fix: Mute people can show Emote Typing Indicator (#2467)

* fix: spacepod keys don't work (#2715)

* FIX: Empty sentinel's plasma storage  (#2727)

* Tweak: SPAI cost (#2719)

* fix: Fix buckled mobs pulling stop (#2678)

* TWEAK: Vox bones fragility (#2711)

* Fix: carbon_lose_organ_signal (#2712)

* Feat: welding masks (#2720)

* Feat: add holoroller (#2707)

* fix: Wands and Staves Open Crates Properly (#2724)

* Fix: additional accesses resets when press "give all access" (#2716)

* Fix: double crates and closets for syndi spawners (#2717)

* Tweak: changes in blood packs on maps, cargo, add separate blood packs for xenos (#2729)

* fix: Fix clickable areas of equipped items (#2642)

* Feat: add new defibrillator, update sprite and targets (#2710)

* Feat: Dirslash (#2713)

* fix: Shadowling Racial Light Damage (#2728)

* Fix: Bad Cables On Taipan (#2714)

* FIX: lattices and catwalks in templates not spawning in late initialization (#2738)

* Fix: jumpboots' infinite flight (#2732)

* Fix: return cure generation (#2739)

* Feat: RCD can now build firelocks (#2718)

* fix: Missing on flashlight sp8 icon (#2750)

* Fix: nodrop clown and mime masks is now actually clown and mime masks. (#2749)

* fix/ref: Finally 37°С! (#2730)

* FEAT: Directed xenomorph spit (#2726)

* feat: icons for slimeman's regrow and toggle color action (#2740)

* Feat: byond 515 initial compat (#2744)

* FEAT: New Typing Emote Indicator (#2721)

* Feat: add fireproof catwalks and lattices (#2746)

* Fix: golden cup sprite. (#2747)

* Fix: dogs not smol anymore. (#2748)

* fix: Cult Memories No More (#2751)

* return our tweaks

* return missed sprites

* Update tgui.bundle.js

* Fix conflict with pulling

---------


* Refactor: Shuttles not take space/lava/chasms with them on transit (#61)

* fix: T-ray scanner rework port from Paradise-EN (#64)

* ref: Deal with magic invisibility value 101

* fix: Port t-ray scanner rework

* ref: Refactor one-letter variables

* changed deadmin rights (#69)

* kinda fix (#70)

* FEAT: Poison pizza from cargo (#55)

* FEAT: Пицца теперь с ананасами

* Update code/datums/supplypacks.dm


* Update code/datums/supplypacks.dm


* Update code/modules/food_and_drinks/food/foods/pizza.dm


* Update crates.dm

* раз уж инитиалайз, значит всю пиццу инитиалайзим

* replace cyanide to slimejelly

* Update code/game/objects/structures/crates_lockers/crates.dm


* Update code/game/objects/structures/crates_lockers/crates.dm


* Update code/game/objects/structures/crates_lockers/crates.dm


* changes for god of changes

* Update code/game/objects/items/devices/pizza_bomb.dm


* recode

---------


* [BOUNTY]Add: donators suits and stuff (#81)

* new suits

* Update colour.dm

* fogot one

* small fixes

* FEAT: Awaking now! (#67)

* feat: Base pixel (#63)

* add: Base pixels!

* feat: Better runechat for offsetted atoms!

* feat: Better pixel shift with big mobs

* Bodycams (#75)

* Feat: portable (onbody) cameras (#1169)

* onbody camera

* moving cameras updates their watchers every movement

* added cameras to security vendors

* Fix: portable camera position update (#1172)

* different fixes

* oops

* Return removed signal

* Feat: auto tag body cams with 'BodyCam' tag (#1326)

* Feat: Автодобавление "BodyCam " к тегу нательных камер.

Стандартный тег Body Camera не меняется
Так как префикс к тэгу занимает 8 символов, то максимальная длина тега сокращена до 24 символов.

* Update code/game/machinery/camera/body_camera.dm



* fix: path for bodycam dmi

---------


* Feat: gun for away missions (#1224) (#76)

* Оружие для гейта

- Добавлена возможность крафта оружия для гейта.
- Крафт оружия без кейса, т.к. оно безопасно.
- Нарисована уникальная иконка.
- Усредненные технологии и стоимость между лазером и АЕГом.
- Оружия в гейте пауков обновили спрайты

Сделано по запросу Максима Wycc220 на Прайм

* Возможность ЕМАГа устройств. Разделение РНД и Гейт-устройства.

* Урезание ЛОРа.

* Внесение правок в описание оружия для гейтеров.

* Update code/modules/projectiles/guns/energy/laser.dm

* Корректировка кода гейт-гана

* Гейт ган, очистка кода по ревью


* Fix: dev VV in runtime menu (#79)

* Feat: Simple healthdolls (TG port) (#83)

* add: filters

* feat: Simple mobs now have healthdolls!

* add: Filters!

* Fix: Alien queen dead sprite (#68)

* tweak: Remove damage from pulling (#71)

* Feat: (Cyberiad) Polarized windows in QM office and a couple of improvements. (#65)

* polarized qm office

* couple improvements

* some fixes

* feat: DNA TTS (#29)

* Revert "Revert "feat: Human TTS is now in DNA, + fixes" (#2599)"

This reverts commit 6e65926150352c430b679c6687dc202964e91493.

* fix: TTS gets from prefs

* ref: Changes to async tts seed update

* ref: Voice changer little changes

* Update code/modules/mob/mob.dm


* ref: mimicing -> mimicking

---------


* Tweak: loneop nerf 2 (#72)

* tmp

* nerf

* typo

* typo

* best covering of equip cases

* even better

* naming fix

* fix

* removed usused

---------

+++ code/game/turfs/turf.dm
@@ -188,7 +188,7 @@
 // Removes all signs of lattice on the pos of the turf -Donkieyo
 /turf/proc/RemoveLattice()
 	var/obj/structure/lattice/L = locate(/obj/structure/lattice, src)
-	if(L)
+	if(L && L.initialized)
 		qdel(L)
 
 /turf/proc/dismantle_wall(devastated = FALSE, explode = FALSE)
@@ -434,7 +434,7 @@
 		for(var/obj/O in contents) //this is for deleting things like wires contained in the turf
 			if(O.level != 1)
 				continue
-			if(O.invisibility == INVISIBILITY_MAXIMUM)
+			if(O.invisibility == INVISIBILITY_MAXIMUM || O.invisibility == INVISIBILITY_ABSTRACT)
 				O.singularity_act()
 	ChangeTurf(baseturf)
 	return 2
